chore: move toml function to config package and improve comments
